+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonForm.cs
@@ -350,7 +350,7 @@ public class KryptonForm : VisualForm,
 
     private Color GetGripDotColor()
     {
-        var palette = GetResolvedPalette();
+        var palette = GetResolvedPalette() ?? KryptonManager.CurrentGlobalPalette;
         Color back = palette.GetBackColor1(PaletteBackStyle.FormMain, PaletteState.Normal);
         if (back == GlobalStaticValues.EMPTY_COLOR || back.IsEmpty)
         {
@@ -401,7 +401,7 @@ public class KryptonForm : VisualForm,
     {
         // First, ask the palette for a themed sizing grip image (RTL-aware)
         var isRtl = RightToLeftLayout ? RightToLeft.Yes : RightToLeft.No;
-        Image? themedGrip = GetResolvedPalette().GetSizeGripImage(isRtl);
+        Image? themedGrip = (GetResolvedPalette() ?? KryptonManager.CurrentGlobalPalette).GetSizeGripImage(isRtl);
         if (themedGrip is null)
         {
             return false;
@@ -1207,7 +1207,7 @@ public class KryptonForm : VisualForm,
     /// Create the redirector instance.
     /// </summary>
     /// <returns>PaletteRedirect derived class.</returns>
-    protected override PaletteRedirect CreateRedirector() => new FormPaletteRedirect(GetResolvedPalette(), this);
+    protected override PaletteRedirect CreateRedirector() => new FormPaletteRedirect(GetResolvedPalette() ?? KryptonManager.CurrentGlobalPalette, this);
 
     internal class FormPaletteRedirect : PaletteRedirect
     {
@@ -1592,11 +1592,12 @@ public class KryptonForm : VisualForm,
                 }
             }
         }
-        else if (_internalPanelState == InheritBool.Inherit)
+        else if (_internalPanelState == InheritBool.Inherit && !DesignMode)
         {
             // #2448 | Work-around / fix | Only runs on non mdi containers
             // This happens when KForm is instantiated manually without designer source,
             // which runs the layout methods through InitializeComponent.
+            // This block is only to be executed at runtime.
             SuspendLayout();
             ResumeLayout(false);
         }
@@ -2232,7 +2233,7 @@ public class KryptonForm : VisualForm,
             // Decide if we should have custom chrome applied
             var needChrome = UseThemeFormChromeBorderWidth &&
                              KryptonManager.UseThemeFormChromeBorderWidth &&
-                             (GetResolvedPalette().UseThemeFormChromeBorderWidth == InheritBool.True);
+                             ((GetResolvedPalette() ?? KryptonManager.CurrentGlobalPalette).UseThemeFormChromeBorderWidth == InheritBool.True);
 
             // Is there a change in custom chrome requirement?
             if (UseThemeFormChromeBorderWidth != needChrome